ada v0.1.0 Ada.Backup.Strategy behaviour

The Ada.Backup.Strategy behaviour defines a module capable of backing up and restoring files from a specific provider.

Types

contents()

contents() :: binary()

The contents of the backup file

name()

name() :: String.t()

The name to use when saving the file

path()

path() :: String.t()

The path where a backup is stored

Callbacks

configured?()

configured?() :: boolean()

This callback should check if the module is properly setup, e.g. if any access token is present.

The function will be invoked at application boot. If it returns false, backups will be disabled.

download_file(path)

download_file(path()) :: {:ok, contents()} | {:error, term()}

Download a file at a given path, returning its contents.

list_files()

list_files() :: {:ok, [path()]} | {:error, term()}

Returns a list of paths where backups are stored. Each one of these paths should be compatible with the download_file/1 callback.

upload_file(name, contents)

upload_file(name(), contents()) :: {:ok, path()} | {:error, term()}

Uploads a file with the specified contents under the given name, returning its path.