ada v0.1.0 Ada.Backup.Uploader

Controls the upload of the database file to the configured backup location.

Types

start_opts()

start_opts() :: [{:strategy, module()}]

Functions

child_spec(init_arg)

Returns a specification to start this module under a supervisor.

See Supervisor.

save_now()

save_now() :: {:ok, Ada.Backup.Strategy.path()} | no_return()

Saves a copy of the database, naming it with the current timestamp.

save_today()

save_today() :: {:ok, Ada.Backup.Strategy.path()} | no_return()

Saves a copy of the database, naming it with today's date.

start_link(opts)

start_link(start_opts()) :: GenServer.on_start()

Starts the uploaded process. Requires a strategy.