ada v0.1.0 Ada.CLI.Helpers

Provides helper functions to ease the composition of CLI tasks.

Functions

connect!(cli_node, target_node)

Uses the erlang distribution to connect the CLI node to the device node.

ensure_location_dependencies!()

Only working on Mac, checks for the dependencies needed to infer the current location of the machine from the command line.

get_current_location_data()

Only working on Mac, returns the current location data.