ada v0.1.0 Ada.CLI.Markup

This module defines semantic helpers that can be used to format CLI-based reports.

It includes (among other things) titles, headings, paragraphs (with support for wrapping text), lists and bars.

The recommended usage pattern is to build lists of elements and then pass them to IO.iodata_to_binary/1 for conversion to a printable binary.

Functions

bar(length)

Returns a bar of the specified length.

break()

New line separator

dash()

Returns a dash, followed by a space

double_title(left, right, max_length)

Separates left text and right text with a continuous line, pushing them to the left and right border of the page. The left text is wrapped in red.

E.g. Left ———————————————————————————————— Right

ellipsis(string, max_length)

Truncates the text at the specified length, appending an ellipsis

emdash()

Returns an emphasis dash, wrapped in spaces

h1(contents)

Returns the text wrappend on the left, in primary color and adds a break at the end.

h2(contents)

Returns the text wrappend on the left, in secondary color and adds a break at the end.

h3(contents)

Returns the text wrappend on the left, in tertiary color and adds a break at the end.

left_pad()

Left padding

list_item(name, values)

Returns a list item, i.e. a definition with a name (wrapped in secondary colour) and a value.

Values are pretty printed according to their type:

maps, keyword lists and lists of tuples get expanded one pair per line, with the pair elements separated by a :

lists get expanded one element per line

other values are printed on one line beside the name.

p(contents, max_length)

Returns the text, wrapped at the specified length.

All lines are padded on the left and it adds a break at the end.

primary(contents)

Wraps contents in primary color

secondary(contents)

Wraps contents in secondary color

space()

Space separator

tertiary(contents)

Wraps contents in tertiary color

title(contents)

Returns the text padded on the left, in red color and adds a break at the end.