ada v0.1.0 Ada.Display

Lower level driver to directly control the contents of the LED screen.

Types

content()

content() :: static_spec() | cycle_spec()

cycle_spec()

cycle_spec() :: {:cycle, [{Ada.Display.Driver.buffer(), duration_ms()}]}

Represents a specification for a cyclic screen, which loops over different buffers, each one with a specific duration.

For example:

[
  {Ada.UI.Helpers.chars_to_matrix('A   '), 200},
  {Ada.UI.Helpers.chars_to_matrix(' A  '), 200},
  {Ada.UI.Helpers.chars_to_matrix('  A '), 200},
  {Ada.UI.Helpers.chars_to_matrix('   A'), 200}
]

duration_ms()

duration_ms() :: pos_integer()

static_spec()

static_spec() :: {:static, Ada.Display.Driver.buffer()}

Represents a specification for a static screen, constantly displaying the same contents.

For example:

{:static, Ada.UI.Helpers.chars_to_matrix('1234')}

Functions

get_brightness()

get_brightness() :: Ada.Display.Driver.brightness()

Gets the screen brigthness.

set_brightness(brightness)

set_brightness(Ada.Display.Driver.brightness()) :: :ok | no_return()

Sets the screen brigthness.

set_content(content)

set_content(content()) :: :ok | {:error, :cycle_spec_too_short} | no_return()

Sets the screen contents.

Please see static_spec/0 and cycle_spec/0 for information about supported formats.

turn_off()

turn_off() :: :ok

Turn off the screen.

turn_on()

turn_on() :: :ok

Turn on the screen.