ada v0.1.0 Ada.Display.Driver behaviour

This behaviour represents a generic display driver, which needs to be implemented when supporting a new LED display or similar.

Types

brightness()

brightness() :: 1..255

Represents the brigthness of the display, from low (1) to high (255).

buffer()

buffer() :: Matrix.matrix()

A data structure representing the contents of the display.

Callbacks

default_content()

default_content() :: buffer()

Returns the buffer that should be set when the display is turned on.

set_brightness(brightness)

set_brightness(brightness()) :: :ok

Synchronously set the current brightness, so that the display can be updated.

set_buffer(buffer)

set_buffer(buffer()) :: :ok

Synchronously set the current buffer, so that it can be displayed.

set_default_brightness()

set_default_brightness() :: :ok

Sets the default brightness that's used when the display is turned on.