ada v0.1.0 Ada.Preferences

Control system wide preferences.

Functions

all()

all() :: [{Ada.Schema.Preference.name(), Ada.Schema.Preference.value()}]

Returns all existing preferences as tuples.

cast(arg1)

cast(String.t()) ::
  {:ok, Ada.Schema.Preference.name()} | {:error, :invalid_preference_name}

Safely cast a preference name to its atom form.

get(name)

get(Ada.Schema.Preference.name()) :: Ada.Schema.Preference.value()

Returns the value of a given preference.

load_defaults!()

load_defaults!() :: [Ada.Schema.Preference.t()]

Loads default values in the database without overwriting already existing entries.

set(name, value)

set(Ada.Schema.Preference.name(), Ada.Schema.Preference.value()) :: :ok

Sets the value of a given preference.