ada v0.1.0 Ada.Schema.Frequency

The module expresses the idea of something that can be repeated at regular intervals.

While it's used mainly with Ada.Schema.ScheduledTask, it can be applied to other use cases.

See t/0 for details.

Types

t()

t() :: %Ada.Schema.Frequency{
  day_of_week: 1..7,
  hour: 0..23,
  id: nil | String.t(),
  minute: 0..59,
  second: 0..59,
  type: String.t()
}

A frequency is determined by a type (hourly, daily or weekly) and day of the week, hour, minute and second.

Depending on the type, some fields are irrelevant (e.g. minutes for a weekly frequency).

Functions

changeset(frequency, params)

changeset(t(), map()) :: Ecto.Changeset.t()

Returns a changeset, starting from a frequency and a map of attributes to change.

daily?(frequency)

daily?(t()) :: boolean()

Returns true for a daily frequency.

hourly?(frequency)

hourly?(t()) :: boolean()

Returns true for a hourly frequency.

matches_time?(frequency, datetime)

matches_time?(t(), DateTime.t()) :: boolean()

Returns true for a frequency that matches a given datetime, where matching is defined as:

same day of the week, hour and zero minutes and seconds for a weekly frequency

same hour, same minute and zero seconds for a daily frequency

same minute and second for a hourly frequency

weekly?(frequency)

weekly?(t()) :: boolean()

Returns true for a weekly frequency.