ada v0.1.0 Ada.Schema.Preference

A preference is a pair of key, value settings which affect the behaviour of the entire application.

One example is the preferred timezone.

Types

name()

name() :: atom()

t()

t() :: %Ada.Schema.Preference{
  __meta__: term(),
  inserted_at: nil | DateTime.t(),
  name: name(),
  updated_at: nil | DateTime.t(),
  value: value()
}

value()

value() :: nil | String.t()

Functions

changeset(preference, params \\ %{})

changeset(t(), map()) :: Ecto.Changeset.t()

Returns a changeset, starting from a preference and a map of attributes to change.