ada v0.1.0 Ada.Schema.ScheduledTask

Represents a boilerplate for the recurring execution of a workflow.

Captures the workflow to run, its frequency and params.

See t/0 for more details.

Types

t()

t() :: %Ada.Schema.ScheduledTask{
  __meta__: term(),
  frequency: Ada.Schema.Frequency.t(),
  id: String.t(),
  inserted_at: DateTime.t(),
  params: map(),
  transport: Ada.Workflow.transport(),
  updated_at: DateTime.t(),
  version: pos_integer(),
  workflow_name: Ada.Workflow.t()
}

A scheduled task is mainly defined by:

a workflow name, deciding the workflow that needs to be run

a frequency, determining how often the task is run (see Ada.Schema.Frequency)

a map of params, which are going to be passed to the workflow when run

a transport, deciding the transport used to communicate the workflow result to the relevant user

Functions

changeset(scheduled_task, params \\ %{})

changeset(t(), map()) :: Ecto.Changeset.t()

Returns a changeset, starting from a scheduled task and a map of attributes to change.

daily?(scheduled_task)

daily?(t()) :: bool()

Returns true for a daily task.

hourly?(scheduled_task)

hourly?(t()) :: bool()

Returns true for an hourly task.

matches_time?(st, datetime)

matches_time?(t(), DateTime.t()) :: bool()

Returns true for a task that matches a given datetime, where matching is defined as:

same day of the week, hour and zero minutes and seconds for a weekly task

same hour, same minute and zero seconds for a daily task

same minute and second for a hourly task

preview(st, ctx \\ [])

preview(t(), Keyword.t()) :: Ada.Workflow.raw_data_result()

Previews the results of a scheduled task by looking at its raw data.

run(st, ctx \\ [])

run(t(), Keyword.t()) :: Ada.Workflow.run_result()

Runs a scheduled task resolving the contained workflow.

weekly?(scheduled_task)

weekly?(t()) :: bool()

Returns true for a weekly task.