ada v0.1.0 Ada.Schema.User

Represents a system user, identified by a numeric ID.

Fields are defined in combination with workflows (see Ada.Worfklow)

Types

t()

t() :: %Ada.Schema.User{
  __meta__: term(),
  email: String.t(),
  id: String.t(),
  inserted_at: DateTime.t(),
  last_fm_username: Ada.Source.LastFm.username(),
  name: String.t(),
  updated_at: DateTime.t()
}

Functions

changeset(user, params \\ %{})

changeset(t(), map()) :: Ecto.Changeset.t()

Returns a changeset starting from a user and a map of attributes to set.

gravatar_url(user)

gravatar_url(t()) :: String.t()

Given a user, returns their gravatar url.