ada v0.1.0 Ada.Workflow behaviour

The Ada.Workflow module specifies a behaviour which needs to be implemented by all workflows.

Core concepts

A workflow has a set of requirements which define the parameters required for its correct execution (e.g. it may require a user id).

It separates the fetch phase (gathering data) from the format phase (presenting it according to a transport, e.g. email).

From idea to implementation

One may want to fetch the list of trains starting from a specific location and receive them by email.

This translates to a workflow that requires:

a user_id (to resolve the email address to send the email to)

a location_id (to fetch relevant trainline information)

In the fetch phase, the workflow will find user and location in the local repo, then interact with a data source (created separately under the Ada.Source namespace) to retrieve the list of trains.

In the format phase, this list of trains, along with any other data coming from the fetch phase, can be formatted according to the transport.

It's important that all side-effectful operations (db queries, http api interactions, current time, etc.) are performed in the fetch phase. This way the format phase can be completely pure, immutable and easy to test.

All workflow module names need to start with Ada.Worfklow to be correctly resolved by the runtime.

Examples

Please see Ada.Worfklow.SendLastFmReport or any other existing workflow module.

Types

ctx()

ctx() :: Keyword.t()

raw_data()

raw_data() :: term()

raw_data_result()

raw_data_result() :: {:ok, raw_data()} | {:error, term()}

requirements()

requirements() :: %{optional(atom()) => term()}

run_result()

run_result() :: {:ok, Ada.Email.t()} | {:error, term()}

t()

t() :: module()

transport()

transport() :: :email

validation_errors()

validation_errors() :: [{atom(), Ecto.Changeset.error()}]

Functions

normalize_name(workflow_name)

normalize_name(t() | String.t()) :: String.t()

Normalizes a workflow name to string, avoiding issue in the conversion between a module atom and a string.

raw_data(workflow_name, params, ctx)

raw_data(t(), map(), Keyword.t()) :: raw_data_result()

Executes a workflow's fetch phase, returning the resulting raw data.

run(workflow_name, params, transport, ctx)

run(t(), map(), transport(), Keyword.t()) :: run_result()

Runs a workflow given its name, starting params, a choice of transport and supporting context.

Params are validated and formatted data is checked for compatibility with the chosen transport.

transports()

transports() :: [transport()]

Returns all available transports.

valid_name?(workflow_name)

valid_name?(t()) :: boolean()

Validates that the passed module name is actually a workflow.

validate(workflow_name, params)

validate(t(), map()) ::
  {:ok, map()} | {:error, :invalid_params, validation_errors()}

Validates a map of params according to a workflows's requirements specification.

Callbacks

fetch(map, ctx)

fetch(map(), ctx()) :: {:ok, raw_data()} | {:error, term()}

Given some starting params, return data ready to be formatted.

format(raw_data, transport, ctx)

format(raw_data(), transport(), ctx()) :: {:ok, term()} | {:error, term()}

Given some data resulting from a fetch/2 call and a transport, return a result compatible with such a transport.

For example, for a transport with value :email, a {:ok, %Ada.Email{}} needs to be returned for the workflow to complete successfully.

human_name()

human_name() :: String.t()

Returns a human readable workflow name

requirements()

requirements() :: requirements()

A map representing the workflow data requirements, keyed by the parameter name (e.g. user_id) and its type (:string).

Supports all types handled by Ecto, as under the hood it uses Ecto's Changeset functions to cast and validate data. See https://hexdocs.pm/ecto/2.2.9/Ecto.Schema.html#module-primitive-types for a list of available types.