ada v0.1.0 Ada.Email.Quickchart

Minimal DSL and functions to create image charts via https://quickchart.io.

Doesn't support all api options (yet).

Types

chart_type()

chart_type() :: String.t()

data()

data() :: [number()]

dataset()

dataset() :: %{label: label(), data: data()}

dimension()

dimension() :: pos_integer()

label()

label() :: String.t()

t()

t() :: %Ada.Email.Quickchart{
  data: %{labels: [label()], datasets: [dataset()]},
  height: dimension(),
  type: chart_type(),
  width: dimension()
}

Functions

add_dataset(chart, label, data)

add_dataset(t(), label(), data()) :: t()

Given a chart, add a new dataset, identified by its label and data.

iex> Ada.Email.Quickchart.new()
...> |> Ada.Email.Quickchart.add_dataset("Sales", [1,2,3])
%Ada.Email.Quickchart{
  data: %{datasets: [%{label: "Sales", data: [1,2,3]}], labels: []},
  height: 300,
  type: "bar",
  width: 516
}

add_labels(chart, labels)

add_labels(t(), [label()]) :: t()

Given a chart, add general labels.

iex> Ada.Email.Quickchart.new()
...> |> Ada.Email.Quickchart.add_labels(["May", "June"])
%Ada.Email.Quickchart{
  data: %{datasets: [], labels: ["May", "June"]},
  height: 300,
  type: "bar",
  width: 516
}

new(type \\ "bar")

new(chart_type()) :: t()

Given a chart type, return a new chart. Defaults to bar

iex> Ada.Email.Quickchart.new()
%Ada.Email.Quickchart{
  data: %{datasets: [], labels: []},
  height: 300,
  type: "bar",
  width: 516
}
iex> Ada.Email.Quickchart.new("line")
%Ada.Email.Quickchart{
  data: %{datasets: [], labels: []},
  height: 300,
  type: "line",
  width: 516
}

set_dimensions(chart, width, height)

set_dimensions(t(), dimension(), dimension()) :: t()

Given a chart, set its dimensions.

iex> Ada.Email.Quickchart.new()
...> |> Ada.Email.Quickchart.set_dimensions(200, 200)
%Ada.Email.Quickchart{
  data: %{datasets: [], labels: []},
  height: 200,
  type: "bar",
  width: 200
}

to_url(chart)

to_url(t()) :: String.t()

Given a chart, returns its corresponding quickchart url.

iex> Ada.Email.Quickchart.new()
...> |> Ada.Email.Quickchart.add_labels(["April", "May", "June"])
...> |> Ada.Email.Quickchart.add_dataset("Sales", [1,2,3])
...> |> Ada.Email.Quickchart.to_url()
"https://quickchart.io/chart?c=%7B%27data%27%3A%7B%27datasets%27%3A%5B%7B%27data%27%3A%5B1%2C2%2C3%5D%2C%27label%27%3A%27Sales%27%7D%5D%2C%27labels%27%3A%5B%27April%27%2C%27May%27%2C%27June%27%5D%7D%2C%27type%27%3A%27bar%27%7D&height=300&width=516"